# app/Http/Requests/Authentication/LoginRequest.php
<?php

namespace App\Http\Requests\Authentication;

use Illuminate\Foundation\Http\FormRequest;

/** 
 * @OA\Schema()
 */
class LoginRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     * @OA\Property(property="email", type="string", format="email", description="Email address of the user")
     * @OA\Property(property="password", type="string", format="password", description="Password of the user")
     * 
     * @return array<string, mixed>
     */
    public function rules()
    {
        return [
            'email' => 'required|email',
            'password' => 'required',
        ];
    }
}